+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Update the `org.realityforge.guiceyloops` artifact to version `0.113`.
 * Update the java version required to `17`.
 * Remove several typos in `dependencies.yml`
+* Ensure that the resolver will resolve an artifacts based on the set of explicitly listed repositories.
 
 ### [v0.17](https://github.com/realityforge/bazel-depgen/tree/v0.17) (2022-05-24) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.16...v0.17)